ws: Switch to mergerfs
This seems to play a little nicer with chroot-git, given that it
doesn't end up traversing devices from its perspective.

The mergerfs change necessitates moving srcpkgs into a root/ directory
so it can be merged with void-packages without creating a recursive
merge over void-packages itself. This doesn't really change very much.
The same could possibly be done with two mounts, but it's not really
worth it.

- Add `binds create/rm` commands.
- Switch from bindfs to mergerfs.
- Add COPYING text.