Add xtools to PATH for ws run <cmd>
This fetches a copy of xtools and sticks it in the PATH for `ws run`
and other commands to use. This way using, for example, xlint on
packages should be easier than if using the xtools xbps package -- only
really useful for following master, though. May add a flag to disable
this later if it turns out it's a problem.

- Switched back to /bin/sh since there's no longer any need to support
  exported functions (the only reason for the bash switch).

- xargs for each_pkg will now print the command it's running. This is
  just for the sake of making it easier to see what's running and when
  it started.